add a bench; move the implementation into lib.rs
