+++ Cargo.toml
@@ -7,3 +7,11 @@ edition = "2018"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
+
+[lib]
+name = "lc3"
+path = "src/lib.rs"
+
+[[bin]]
+name = "main"
+path = "src/main.rs"
